requirements-prod.txt
-r requirements.txt
gunicorn==21.2.0
python-dotenv==1.0.1
whitenoise==6.6.0
psycopg2-binary==2.9.9  # Para PostgreSQL
